backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Pydantic (Updated for Python 3.13)
pydantic[email]==2.10.4
pydantic-settings==2.6.1

# OpenAI & API
openai==1.59.3
httpx==0.28.1
websockets==14.1

# Async Support
asyncpg==0.30.0
aiosqlite==0.20.0

# Utilities
redis==5.2.1
celery==5.4.0
email-validator==2.2.0

# Development
watchfiles==0.24.0
